# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Copy Trading Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List all webhooks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Create a webhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /webhooks/{id}:
    put:
      operationId: updateWebhook
      tags: [webhooks]
      summary: Update a webhook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUpdate"
      responses:
        "200":
          description: Updated webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /webhooks/{token}/test:
    post:
      operationId: testWebhook
      tags: [webhooks]
      summary: Test a webhook (dry run)
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Webhook:
      type: object
      required: [id, name, token, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        token:
          type: string
        targetAccounts:
          type: array
          items:
            type: integer
        defaultSymbol:
          type: ["string", "null"]
        defaultLeverage:
          type: ["integer", "null"]
        isActive:
          type: boolean
        lastTriggered:
          type: ["string", "null"]
        createdAt:
          type: string
    WebhookInput:
      type: object
      required: [name, targetAccounts]
      properties:
        name:
          type: string
        targetAccounts:
          type: array
          items:
            type: integer
        defaultSymbol:
          type: ["string", "null"]
        defaultLeverage:
          type: ["integer", "null"]
        isActive:
          type: boolean
    WebhookUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        targetAccounts:
          type: array
          items:
            type: integer
        defaultSymbol:
          type: ["string", "null"]
        defaultLeverage:
          type: ["integer", "null"]
        isActive:
          type: ["boolean", "null"]
...